chore: rename to trustbloc DID method
- rename spec references.
- rename BDD references.
- rename vdri package.
- fix CI.
- TODO: rename repository.

Closes: #28
