+++ pkg/internal/mock/selection/mock_service.go
@@ -5,7 +5,7 @@ SPDX-License-Identifier: Apache-2.0
 
 package discovery
 
-import "github.com/trustbloc/bloc-did-method/pkg/vdri/bloc/endpoint"
+import "github.com/trustbloc/bloc-did-method/pkg/vdri/trustbloc/endpoint"
 
 // MockSelectionService implements a mock selection service
 type MockSelectionService struct {
