+++ pkg/vdri/trustbloc/config/consortium_test.go
@@ -23,11 +23,11 @@ var payload = `
 	"stakeholders": [
 		{
 			"domain": "bar.baz",
-			"did": "did:bloc:foo.bar:zQ1234567890987654321"
+			"did": "did:trustbloc:foo.bar:zQ1234567890987654321"
 		},
 		{
 			"domain": "baz.qux",
-			"did": "did:bloc:foo.bar:zQ0987654321234567890"
+			"did": "did:trustbloc:foo.bar:zQ0987654321234567890"
 		}
 	],
 	"previous": ""
