+++ pkg/vdri/trustbloc/endpoint/endpoint.go
