+++ test/bdd/features/did_method_e2e.feature
@@ -9,6 +9,6 @@
 Feature: Using DID method REST API
 
   @e2e
-  Scenario: create bloc did and resolve it through
-    Given Bloc DID is created from domain "testnet.trustbloc.local:80"
+  Scenario: create trustbloc did and resolve it through
+    Given TrustBloc DID is created from domain "testnet.trustbloc.local:80"
     Then Resolving created DID through resolver URL "http://localhost:8080/1.0/identifiers"
