+++ test/bdd/fixtures/did-method-rest/docker-compose.yml
@@ -8,8 +8,8 @@ version: '2'
 
 services:
 
-  bloc.did.method.example.com:
-    container_name: bloc.did.method.example.com
+  trustbloc.did.method.example.com:
+    container_name: trustbloc.did.method.example.com
     image: docker.pkg.github.com/trustbloc/bloc-did-method/did-method-rest:latest
     environment:
       - DID_METHOD_HOST_URL=0.0.0.0:8070
