+++ test/bdd/fixtures/discovery-server/config-data/consortium.json
@@ -1,6 +1,6 @@
 {
   "domain": "consortium",
-  "did": "did:bloc:testnet.trustbloc.local:zQ1234567890987654321",
+  "did": "did:trustbloc:testnet.trustbloc.local:zQ1234567890987654321",
   "policy": {
     "cache": {"max-age": 123456789}
   },
