+++ test/bdd/fixtures/sidetree-mock/docker-compose.yml
@@ -16,7 +16,7 @@ services:
       - SIDETREE_MOCK_TLS_KEY=/etc/sidetree/tls/ec-key.pem
       - SIDETREE_MOCK_HOST=0.0.0.0
       - SIDETREE_MOCK_PORT=48326
-      - SIDETREE_MOCK_DID_NAMESPACE=did:bloc:testnet.trustbloc.local
+      - SIDETREE_MOCK_DID_NAMESPACE=did:trustbloc:testnet.trustbloc.local
     ports:
       - 48326:48326
     volumes:
