+++ test/bdd/fixtures/stakeholder-server/docker-compose.yml
@@ -18,7 +18,7 @@ services:
     ports:
       - "8088:8088"
     volumes:
-      - ./config:/web/.well-known/did-bloc
+      - ./config:/web/.well-known/did-trustbloc
     networks:
       - did-method-rest_bdd_net
