+++ test/bdd/fixtures/universalresolver/config.json
@@ -5,8 +5,8 @@
       "url": "https://ashburn.capybara.veres.one/dids/$1"
     },
     {
-      "pattern": "^(did:bloc:testnet.trustbloc.local:.+)$",
-      "url": "http://bloc.did.method.example.com:8070/resolveDID?did=$1"
+      "pattern": "^(did:trustbloc:testnet.trustbloc.local:.+)$",
+      "url": "http://trustbloc.did.method.example.com:8070/resolveDID?did=$1"
     }
   ]
 }
